fix(SearchAddressResult.kt) : 파일 분리
