+++ app/src/main/java/com/manna/presentation/search/SearchAddressItem.kt
@@ -1,22 +1,6 @@
 package com.manna.presentation.search
 
-import android.os.Parcelable
 import com.manna.network.model.search_address.SearchAddress
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class SearchAddressResult(
-    val addressName: String,
-    val keyWord: String,
-    val placeName: String,
-    val latitude: String,
-    val longitude: String
-): Parcelable {
-    companion object {
-        fun of(item: SearchAddressItem): SearchAddressResult =
-            SearchAddressResult(item.addressName, item.keyWord, item.placeName, item.latitude, item.longitude)
-    }
-}
 
 data class SearchAddressItem(
     val addressName: String,
